完成热更部分可能出现的http网络模块错误兼容；

+++ Assets/Scripts/Util/HttpUtil.cs
@@ -4,82 +4,158 @@ using System.Collections;
 using System.IO;
 using System.Net;
 using System.Text;
+using System.Threading;
 
+public class HttpResult
+{
+    public int code;
+    public string data;
+}
 public class HttpUtil
 {
-    public CookieContainer cookies = new CookieContainer();
     /// <summary>
-    /// GET方法(自动维护cookie)
+    /// 主线程
     /// </summary>
-    public string Get(string url, string referer = "", int timeout = 2000, Encoding encode = null)
+    private SynchronizationContext _mainThreadSynContext;
+
+    public HttpUtil()
     {
-        string dat;
+        _mainThreadSynContext = SynchronizationContext.Current;
+    }
+    private void MainCallBack(object param)
+    {
+        var tag = (string)param.GetType().GetProperty("tag").GetValue(param);
+        switch (tag)
+        {
+            case "Get":
+                var getCb = (Action)param.GetType().GetProperty("cb").GetValue(param);
+                if (getCb != null) getCb();
+                break;
+            case "GetError":
+                var getErrorCb = (Action)param.GetType().GetProperty("cb").GetValue(param);
+                if (getErrorCb != null) getErrorCb();
+                break;
+            case "Post":
+                var postCb = (Action)param.GetType().GetProperty("cb").GetValue(param);
+                if (postCb != null) postCb();
+                break;
+            case "PostError":
+                var postErrorCb = (Action)param.GetType().GetProperty("cb").GetValue(param);
+                if (postErrorCb != null) postErrorCb();
+                break;
+            case "DownloadOver":
+                var downloadOverCb = (Action)param.GetType().GetProperty("cb").GetValue(param);
+                if (downloadOverCb != null) downloadOverCb();
+                break;
+            case "DownloadIng":
+                var size = (int)param.GetType().GetProperty("size").GetValue(param);
+                var count = (int)param.GetType().GetProperty("count").GetValue(param);
+                var downloadIngCb = (Action<int, int>)param.GetType().GetProperty("cb").GetValue(param);
+                if (downloadIngCb != null) downloadIngCb(size, count);
+                break;
+        }
+
+    }
+    public void Get_Asyn(string url, Action<HttpResult> cb = null, Action errorCb = null)
+    {
+        Thread thread = new Thread(new ThreadStart(() =>
+        {
+            HttpResult result = Get(url, () =>
+            {
+                _mainThreadSynContext.Post(new SendOrPostCallback(MainCallBack), new { tag = "GetError", cb = errorCb });
+            });
+            Action t_cb = () => { if (cb != null) cb(result); };
+            _mainThreadSynContext.Post(new SendOrPostCallback(MainCallBack), new { tag = "Get", cb = t_cb });
+        }));
+        thread.Start();
+    }
+
+    /// <summary>
+    /// GET方法
+    /// </summary>
+    public HttpResult Get(string url, Action errorCb = null)
+    {
         HttpWebResponse res = null;
         HttpWebRequest req = null;
         try
         {
             req = (HttpWebRequest)WebRequest.Create(url);
-            req.CookieContainer = cookies;
             req.AllowAutoRedirect = false;
-            req.Timeout = timeout;
-            req.Referer = referer;
-            req.UserAgent = "Mozilla/5.0 (Windows NT 10.0;%20WOW64; rv:47.0) Gecko/20100101 Firefox/47.0";
+            req.Timeout = 1000;
             res = (HttpWebResponse)req.GetResponse();
-            cookies.Add(res.Cookies);
-            dat = new StreamReader(res.GetResponseStream(), encode ?? Encoding.UTF8).ReadToEnd();
+            StreamReader sr = new StreamReader(res.GetResponseStream(), Encoding.Default);
+            int code = (int)res.StatusCode;
+            string data = sr.ReadToEnd();
+            sr.Close();
             res.Close();
             req.Abort();
+            return new HttpResult() { code = code, data = data };
         }
         catch
         {
-            return null;
+            if (errorCb != null) errorCb();
+            return new HttpResult() { code = -1, data = null };
         }
-        return dat;
+    }
+
+    public void Post_Asyn(string url, string body, Action<HttpResult> cb = null, Action errorCb = null)
+    {
+        Thread thread = new Thread(new ThreadStart(() =>
+        {
+            HttpResult result = Post(url, body, () =>
+            {
+                _mainThreadSynContext.Post(new SendOrPostCallback(MainCallBack), new { tag = "PostError", cb = errorCb });
+            });
+            Action t_cb = () => { if (cb != null) cb(result); };
+            _mainThreadSynContext.Post(new SendOrPostCallback(MainCallBack), new { tag = "Post", cb = t_cb });
+        }));
+        thread.Start();
     }
 
 
     /// <summary>
-    /// Post方法(自动维护cookie)
+    /// Post方法
     /// </summary>
-    public string Post(string url, string postdata, CookieContainer cookie = null, string referer = "", int timeout = 2000, Encoding encode = null)
+    public HttpResult Post(string url, string body, Action errorCb = null)
     {
-        string html = null;
-        HttpWebRequest request;
-        HttpWebResponse response;
-        if (encode == null) encode = Encoding.UTF8;
+        HttpWebResponse res;
+        HttpWebRequest req;
+        Encoding encode = Encoding.Default;
         try
         {
-            byte[] byteArray = encode.GetBytes(postdata); // 转化
-            request = (HttpWebRequest)WebRequest.Create(new Uri(url));
-            if (cookie == null) cookie = new CookieContainer();
-            request.CookieContainer = cookie;
-            request.UserAgent = "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.2; SV1; .NET CLR 1.1.4322; InfoPath.1)";
-            request.Method = "POST";
-            request.Referer = referer;
-            request.ContentType = "application/x-www-form-urlencoded";
-            request.ContentLength = byteArray.Length;
-            request.Timeout = timeout;
-            Stream newStream = request.GetRequestStream();
+            byte[] byteArray = encode.GetBytes(body);
+            req = (HttpWebRequest)WebRequest.Create(new Uri(url));
+            req.Method = "POST";
+            req.Accept = "*/*";
+            req.UserAgent = "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.2; SV1; .NET CLR 1.1.4322; InfoPath.1)";
+            req.ContentType = "text/plain";
+            req.ContentLength = byteArray.Length;
+            req.Timeout = 1000;
+            Stream newStream = req.GetRequestStream();
             newStream.Write(byteArray, 0, byteArray.Length);    //写入参数
             newStream.Close();
-            response = (HttpWebResponse)request.GetResponse();
-            cookie.Add(response.Cookies);
-            StreamReader str = new StreamReader(response.GetResponseStream(), encode);
-            html = str.ReadToEnd();
+            res = (HttpWebResponse)req.GetResponse();
+            StreamReader sr = new StreamReader(res.GetResponseStream(), encode);
+            int code = (int)res.StatusCode;
+            string data = sr.ReadToEnd();
+            sr.Close();
+            res.Close();
+            req.Abort();
+            return new HttpResult() { code = code, data = data };
         }
         catch
         {
-            return "";
+            if (errorCb != null) errorCb();
+            return new HttpResult() { code = -1, data = null };
         }
-        return html;
     }
 
-
-    public void Download(string url, string saveFile, string tempFileName = null, Action cb = null, Action<int, int> downloadingCb = null)
+    public void Download(string url, string saveFile, string tempFileName = null, Action cb = null, Action<int, int> downloadingCb = null, Action errorCb = null)
     {
-        MonoSingleton.Instance.StartCoroutine(_download(url, saveFile, tempFileName, cb, downloadingCb));
+        MonoSingleton.Instance.StartCoroutine(new CatchableEnumerator(_download(url, saveFile, tempFileName, cb, downloadingCb, errorCb), errorCb));
+
     }
-    IEnumerator _download(string url, string saveFile, string tempFileName = null, Action cb = null, Action<int, int> downloadingCb = null)
+    IEnumerator _download(string url, string saveFile, string tempFileName = null, Action cb = null, Action<int, int> downloadingCb = null, Action errorCb = null)
     {
         string fileName = Path.GetFileName(saveFile);
         tempFileName = tempFileName == null ? (fileName + ".temp") : tempFileName;
@@ -93,7 +169,6 @@ public class HttpUtil
         }
 
         //打开上次下载的文件或新建文件 
-
         long lStartPos = 0;
         System.IO.FileStream fs;
         if (File.Exists(strTmpFile))
@@ -110,13 +185,18 @@ public class HttpUtil
 
         //打开网络连接 
         HttpWebRequest request;
+        HttpWebResponse response;
+        request = (HttpWebRequest)WebRequest.Create(new Uri(url));
+        request.Method = "HEAD";
+        request.Timeout = 100;
+        response = (HttpWebResponse)request.GetResponse();
+
         request = (System.Net.HttpWebRequest)System.Net.HttpWebRequest.Create(url);
-        request.Timeout = 5000;
+        request.Timeout = 2000;
         if (lStartPos > 0)
         {
             request.AddRange((int)lStartPos); //设置Range值
         }
-
         //获取临时文件大小，注意，不要再创建request，连续请求可能卡死，开启下次请求前请先结束上一次的请求。
         //重要的事说三遍：不要连续请求！不要连续请求！不要连续请求！
         long countLength = request.GetResponse().ContentLength;
@@ -127,33 +207,41 @@ public class HttpUtil
         byte[] nbytes = new byte[len];
         int nReadSize = 0;
         nReadSize = ns.Read(nbytes, 0, len);
+        int size = 0, count = 0;
         while (nReadSize > 0)
         {
             fs.Write(nbytes, 0, nReadSize);
             nReadSize = ns.Read(nbytes, 0, len);
-            int dDownloadedLength = (int)(fs.Length - lStartPos);
-            int dTotalLength = (int)countLength;
-            //UnityEngine.Debug.Log(string.Format("已下载 {0} / {1}", dDownloadedLength, dTotalLength));
-            if (downloadingCb != null) downloadingCb(dDownloadedLength, dTotalLength);
-            yield return false;
+            size = (int)(fs.Length - lStartPos);
+            count = (int)countLength;
+            //UnityEngine.Debug.Log(string.Format("已下载 {0} / {1}", size, count));
+            if (downloadingCb != null) downloadingCb(size + (int)lStartPos, count + (int)lStartPos);
+            yield return 1;
         }
         fs.Close();
         ns.Close();
-        //保存文件夹不存在则创建
-        if (!File.Exists(Path.GetDirectoryName(saveFile)))
+        if (size < count)
         {
-            Directory.CreateDirectory(Path.GetDirectoryName(saveFile));
+            if (errorCb != null) errorCb();
         }
-        //清理原有文件
-        if (File.Exists(saveFile))
+        else
         {
-            File.Delete(saveFile);
-        }
-        //下载完成移动到保存路径
+            //保存文件夹不存在则创建
+            if (!File.Exists(Path.GetDirectoryName(saveFile)))
+            {
+                Directory.CreateDirectory(Path.GetDirectoryName(saveFile));
+            }
+            //清理原有文件
+            if (File.Exists(saveFile))
+            {
+                File.Delete(saveFile);
+            }
+            //下载完成移动到保存路径
 
-        Util.File.MoveTo(strTmpFile, saveFile);
-        if (cb != null) cb();
+            Util.File.MoveTo(strTmpFile, saveFile);
+            if (cb != null) cb();
+        }
+
     }
-
 }
